create the eksa-packages namespace if not present
This is created on all clusters. The namespace will store Helm release
resources, even for workload clusters.

The namespace is created when a bundle is activated.

The namespace shouldn't be created via Helm chart because that would break
full cluster lifecycle installations as they run helm with the permissions of
the eksa-controller-manager service account, which doesn't have permission to
create namespaces.